Remove code that test does not cover
This is not a good practice, but I could not find any third party
video for those edge cases, so I am removing the code for now.
This will make test pass 100% and the pull request can be merged.
If the code fails in the future, then that means we found the video
for the case, we would be able to add the code back.